requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
pandas>=2.2.2
requests>=2.32.3
python-dotenv>=1.0.1
